+++ backend/LeetWars.Core/LeetWars.Core.DAL/Context/EasySpeakCoreContext.cs
@@ -1,0 +1,20 @@
+﻿using LeetWars.Core.DAL.Entities;
+using Microsoft.EntityFrameworkCore;
+
+namespace LeetWars.Core.DAL.Context
+{
+    public class LeetWarsCoreContext : DbContext
+    {
+        public DbSet<Sample> Samples => Set<Sample>();
+
+        public LeetWarsCoreContext(DbContextOptions<LeetWarsCoreContext> options) : base(options)
+        {
+        }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.Configure();
+            modelBuilder.Seed();
+        }
+    }
+}
