--- backend/LeetWars.Core/LeetWars.Core.DAL/Context/EntityConfigurations/LanguageVersionConfig.cs
@@ -1,17 +1,0 @@
-﻿using LeetWars.Core.DAL.Entities;
-using LeetWars.Core.DAL.Entities.HelperEntities;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using System.ComponentModel.DataAnnotations;
-
-namespace LeetWars.Core.DAL.Context.EntityConfigurations
-{
-    public class LanguageVersionConfig : IEntityTypeConfiguration<LanguageVersion>
-    {
-        public void Configure(EntityTypeBuilder<LanguageVersion> builder)
-        {
-            builder.Property(e => e.Version)
-                .HasMaxLength(EntitySettings.MaxShortNameLength);
-        }
-    }
-}
