+++ backend/LeetWars.Core/LeetWars.Core.DAL/Context/EntityConfigurations/SampleConfig.cs
@@ -1,0 +1,17 @@
+﻿using LeetWars.Core.DAL.Entities;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace LeetWars.Core.DAL.Context
+{
+    public class SampleConfig : IEntityTypeConfiguration<Sample>
+    {
+        public void Configure(EntityTypeBuilder<Sample> builder)
+        {
+            builder.Property(e => e.Title)
+             .IsRequired()
+             .HasMaxLength(50);
+
+        }
+    }
+}
