+++ backend/LeetWars.Core/LeetWars.Core.DAL/Entities/Challenge.cs
@@ -6,10 +6,6 @@
         public string Title { get; set; }
         public string Instructions { get; set; }
         public int LevelId { get; set; }
-        public ChallengeLevel? Level { get; set; }
-        public User? Author { get; set; }
-        public List<Tag> Tags { get; } = new();
-        public List<ChallengeVersion> Versions { get; } = new();
 
         public Challenge(string title, string instructions)
         {
