+++ backend/LeetWars.Core/LeetWars.Core.DAL/Entities/ChallengeVersion.cs
@@ -9,11 +9,6 @@ namespace LeetWars.Core.DAL.Entities
         public string InitialSolution { get; set; }
         public string CompleteSolution { get; set; }
         public ChallengeStatus Status { get; set; }
-        public List<LanguageVersion> LanguageVersions { get; } = new();
-        public List<UserSolution> Solutions { get; } = new();
-        public List<Test> Tests { get; } = new();
-        public Language? Language { get; set; }
-
 
         public ChallengeVersion(string initialSolution, string completeSolution)
         {
