--- backend/LeetWars.Core/LeetWars.Core.DAL/Entities/HelperEntities/ChallengeVersionLanguageVersion.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace LeetWars.Core.DAL.Entities.HelperEntities
-{
-    public class ChallengeVersionLanguageVersion
-    {
-        public long ChallengeVersionId { get; set; }
-        public int LanguageVersionId { get; set; }
-    }
-}
