+++ backend/LeetWars.Core/LeetWars.Core.DAL/Entities/Language.cs
@@ -3,8 +3,6 @@
     public class Language : Entity<int>
     {
         public string Name { get; set; }
-        public List<User> Users { get; } = new();
-        public List<LanguageVersion> LanguageVersions { get; } = new();
 
         public Language(string name)
         {
