--- backend/LeetWars.Core/LeetWars.Core.DAL/Entities/Sample.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace LeetWars.Core.DAL.Entities
-{
-    public class Sample : AuditEntity<long>
-    {
-        public string Title { get; set; } = string.Empty;
-        public string? Body { get; set; }
-    }   
-}
