+++ backend/LeetWars.Core/LeetWars.Core.DAL/Entities/Subscription.cs
@@ -11,7 +11,6 @@
         public DateTime SubscribedDate { get; set; }
         public DateTime? UnsubscribedDate { get; set; }
         public bool IsActive { get; set; }
-        public SubscriptionType? SubscriptionType { get; set; }
 
         public Subscription(string paymentSubscriptionId)
         {
