+++ backend/LeetWars.Core/LeetWars.Core.DAL/Entities/Tag.cs
@@ -3,8 +3,6 @@
     public class Tag : Entity<int>
     {
         public string Name { get; set; }
-        public List<Challenge> Challenges { get; } = new();
-
 
         public Tag(string name)
         {
