--- backend/LeetWars.Core/LeetWars.Core.DAL/Entities/UserLanguageLevel.cs
@@ -1,12 +1,0 @@
-﻿using LeetWars.Core.DAL.Enums;
-
-namespace LeetWars.Core.DAL.Entities
-{
-    public class UserLanguageLevel
-    {
-        public long UserId { get; set; }
-        public int LanguageId { get; set; }
-        public LanguageLevel Level { get; set; }
-        public Language? Language { get; set; }
-    }
-}
