--- backend/LeetWars.Core/LeetWars.Core.DAL/Extensions/LimitLengthExtension.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace LeetWars.Core.DAL.Extensions
-{
-    public static class LimitLengthExtension
-    {
-        public static string LimitLength(this string source, int maxLength)
-        {
-            if (source.Length <= maxLength)
-            {
-                return source;
-            }
-
-            return source.Substring(0, maxLength);
-        }
-
-    }
-}
